% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R
\name{plotfit}
\alias{plotfit}
\title{Plot data with fitted line}
\usage{
plotfit(concentration, resistance)
}
\arguments{
\item{concentration}{numeric vector of concentration}

\item{resistance}{numeric vector of resistance}
}
\description{
Plot the linear dependence between the concentration and conductivity of a dilute electrolyte
}
\examples{
concentration <- c(1.e-3, 2.5e-3, 5.e-3, 7.5e-3, 1.e-2)
resistance <- c(5.31, 2.66, 1.40, 0.92, 0.78)
plotfit(concentration, resistance)
}
